[#371] Fix eslint errors (with @kardan)
