+++ client/src/components/dataset/sidebars/ChangeDataType.jsx
@@ -3,87 +3,7 @@ import Immutable from 'immutable';
 import SelectMenu from '../../common/SelectMenu';
 import SidebarHeader from './SidebarHeader';
 import SidebarControls from './SidebarControls';
-
-const parseFormatOptions = [
-  {
-    value: 'YYYY-MM-DD',
-    label: 'YYYY-MM-DD',
-  },
-  {
-    value: 'DD-MM-YYYY',
-    label: 'DD-MM-YYYY',
-  },
-  {
-    value: 'MM-DD-YYYY',
-    label: 'MM-DD-YYYY',
-  },
-  {
-    value: 'CUSTOM',
-    label: 'Custom',
-  },
-];
-
-
-class DateFormatSelect extends Component {
-
-  constructor() {
-    super();
-    this.state = {
-      parseFormat: 'YYYY-MM-DD',
-      customParseFormat: null
-    };
-  }
-
-  onSelectMenuChange(parseFormat) {
-    if (parseFormat === 'CUSTOM') {
-      this.setState({
-        customParseFormat: '',
-        parseFormat: 'CUSTOM'});
-    } else {
-      this.setState({
-        customParseFormat: null,
-        parseFormat,
-      });
-    }
-    this.props.onChange(parseFormat === 'CUSTOM' ? '' : parseFormat)
-  }
-
-  onCustomParseFormatChange(customParseFormat) {
-    this.setState({ customParseFormat });
-    this.props.onChange(customParseFormat);
-  }
-
-  render () {
-    const { parseFormat, customParseFormat } = this.state;
-
-    return (
-      <div className="inputGroup">
-        <label htmlFor="parseFormatMenu">
-          Date format:
-        </label>
-        <SelectMenu
-          name="parseFormatMenu"
-          value={parseFormat}
-          options={parseFormatOptions}
-          onChange={format => this.onSelectMenuChange(format)}
-        />
-      {customParseFormat != null &&
-        <input
-          value={customParseFormat}
-          type="text"
-          className="customParseFormatInput"
-          onChange={evt => this.onCustomParseFormatChange(evt.target.value)}
-        />
-      }
-      </div>
-    );
-  };
-}
-
-
-DateFormatSelect.propTypes = {
-  onChange: PropTypes.func.isRequired,
-};
+import DateFormatSelect from './DateFormatSelect';
 
 function DefaultValueInput({ defaultValue, onChange, newType }) {
   return (
@@ -252,7 +172,7 @@ export default class ChangeDataType extends Component {
             /> : null}
         </div>
         <SidebarControls
-          onApply={ this.isValidTransformation() ? () => onApply(transformation) : null }
+          onApply={this.isValidTransformation() ? () => onApply(transformation) : null}
           onClose={onClose}
         />
       </div>
